Validot init commit
Why so huge commit that contains everything and effectively hides history?

Let me tell you the story:
- at first, Validot had been planned as a small patch to previous project, CoreValidation
- it reached incompatibility very early, for a while it was considered as a new version with breaking changes
- crucial parts were completely rearchitectured and rewritten - at this point it shares very little code with his predecessor

Ultimately, the history of changes would lead you to the code base that is terribly out of date in both idea and implementation.

So I decided to extract the latest commit and do a clean start.