Blocked selecting more than one level down in Member command
It's funny, because I discovered that it's possible at all during the final functional tests.
So I explicitly blocked this as the code, because there are many corner cases (mostly null-related) that I'd like to handle before releasing such a feature.

Anyway, the support for such a case like Member(m => m.NestedMember.MoreNestedMember) will be added later.