Skipping paths with errors without messages
Oh, that was one interesting corner case.
MessagesService calculated capacity correctly, but even if for the path it was zero, it was still creating an entry in the results for it.
The path has no messages (so it was still correct...), but totally redundant.

Also, added tests for this case.