Made objects optional in reference loop detection tests
Ultimately it doesn't make any difference when it comes to the tests' value, but allows to use same test cases for IsValidValidationContext.
Why? For a ValidationContext it doesn't matter - if it's optional, the errors related to null are not going to be added to the collection... but these tests are not about adding errors.
These scenarios are about detecting loops, so making everything optional allows IsValidValidationContext to reach the loop at all (otherwise it stops at the first null member).